Remove ui from build & improve ssh.
- The experimental UI has been removed from build but kept in source by
  prefixing the ui directory with _.
- The ssh config file functionality has been expanded to allow for
  complex ssh configurations that contain pre-existing ProxyCommand
configurations. The two scenarios accounted for are:

1. Lines starting with 'ProxyCommand nc' are considered to be required
when proxies are off and so are commented out when proxies are on, and
uncommented when proxies are off.
2. Lines starting with neither 'ProxyCommand nc' or 'ProxyCommand ssh'
are considered to be more important than any other ProxyCommand lines
and when they are encountered, no change is made to the containing
stanza.